refactor(web): consolidate formatSize + file-icon helpers to shared module
Addresses review issue #89. formatSize was duplicated across files/page.tsx,
file-viewer-panel.tsx, and channel-files-panel.tsx with divergent
zero-byte handling:
- files/page.tsx: !bytes → "" (treated 0 as empty)
- file-viewer-panel.tsx: !bytes && bytes !== 0 → "" (allowed 0)
- channel-files-panel.tsx: bytes == null → "" (allowed 0)

Divergent copies = the exact surface where dedup bugs live. FILE_ICONS
was also split between files/page.tsx (22 entries) and a 12-entry subset
in channel-files-panel.tsx as iconForName, so adding a new extension
required edits in two places plus a third lookup in file-viewer-panel's
typeBadge.

Extract to apps/web/shared/file-display.ts:
- FILE_ICONS: single source of truth, extended with html/htm
- formatSize(bytes?: number | null): explicitly allows 0, treats null and
  undefined as "" so callers see consistent behavior
- getFileIcon(name): extension lookup with 📄 fallback

All three callsites updated. No behavioral change beyond formatSize now
rendering "0 B" consistently instead of "" in the files list.
